Use size_t instead of u_int64_t
git-svn-id: https://www.ibr.cs.tu-bs.de/svn/emma/trunk/src/ibrdtn@21988 1cffb75b-d306-6187-5d55-5a277715007a

+++ include/ibrdtn/data/Block.h
@@ -37,7 +37,7 @@ namespace dtn
 			size_t getSize() const;
 			blob::BLOBReference getBLOBReference();
 
-			u_int64_t _procflags;
+			size_t _procflags;
 
 			virtual void read() {}
 			virtual void commit() {};
